fix: Splitted Bypassere and Key classes
