+++ src/lib/key.d.ts
@@ -1,0 +1,13 @@
+declare class Key {
+  constructor(key: string);
+
+  key: string;
+  valaue: boolean;
+
+  true(): void;
+  false(): void;
+  getKey(): string;
+  getValue(): boolean;
+}
+
+export = Key;
